+++ src/index.js
@@ -1,8 +1,4 @@
-import VueForm from './plugin'
-
 import Form from './Form'
-import FormErrors from './FormErrors'
-
 import HasError from './components/HasError'
 import HasError4 from './components/HasError4'
 import AlertError from './components/Error'
@@ -13,10 +9,8 @@ export {
   Form,
   HasError,
   HasError4,
-  FormErrors,
   AlertError,
   AlertErrors,
-  AlertSuccess
+  AlertSuccess,
+  Form as default
 }
-
-export default VueForm